+++ src/app/banner-svg.tsx
