+++ src/app/dashboard/performance-comparison.tsx
